added the finding of the optimal solution
added the feature that the program stops looking after its found an optimal solution (an order where no person passes a book to the same person twice). something noteworthy is that the optimum solution is only possible for an even number of people and one (potentially the only) solution  is sending your book to your person number + the round number (modulus the number of people to keep it in range).