made the search parallelized
all searches are now done in parallel with 4 threads and the recursions are pretty evenly divided between them. due to the way arc<mutex> works each recursion requires a clone so i dont know if this is actually any faster. using a static shared mutex would probably work better but that requires unsafe code and im not sure if it would cause any race conditions.